% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieving_functions.R
\name{fuel_build}
\alias{fuel_build}
\title{A wrapper for Thomas J. Leeper's MPG package merging data from fueleconomy.gov & the myMPGwebsite}
\usage{
fuel_build(vinid)
}
\arguments{
\item{vin}{VINid number last 5 digits of a VIN}
}
\description{
Retrieve vehicle fuel economy by vehicle ID & Retrieve average fuel economy from My MPG by vehicle ID in dataframe format merged
}
\examples{
fuel_build(18697)
}
\keyword{vin}
