% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieving_functions.R
\name{vin_batcher}
\alias{vin_batcher}
\title{vin_batcher: outputs a vector of vins for batch upload to NHTSA}
\usage{
vin_batcher(vinlist)
}
\arguments{
\item{vinlist}{a vector of VIN numbers (list or dataframe column)}
}
\description{
Input a list of VINS to return a vector semicolon separated
}
\examples{
vin_batcher(c("1G2HX54K724118697","1G2HX54K724118696")) #save as csv and copy paste values into: https://vpic.nhtsa.dot.gov/api/
}
\keyword{vin}

