% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieving_functions.R
\name{vin_dieselex}
\alias{vin_dieselex}
\title{Return EXTENDED VIN information from the NHTSA database}
\usage{
vin_dieselex(vin, sec = 5, tidyup = TRUE)
}
\arguments{
\item{vin}{vin number in quotes}

\item{sec}{denotes Sys.sleep() between request, Defaults to 5 seconds}

\item{tidyup}{simplifies json response into a dataframe, Defaults to TRUE}
}
\description{
vin_dieselex breaks into NHTSA's database to quickly and most furiously return VIN information. More information on the NHTSA API is available at https://vpic.nhtsa.dot.gov/api/#
}
\examples{
vin_dieselex("1G2HX54K724118697", sec = 1, tidyup = TRUE)
}
\keyword{vin}

