% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieving_functions.R
\name{vin_dieseltype}
\alias{vin_dieseltype}
\title{Return vehcile type by id from the NHTSA database}
\usage{
vin_dieseltype(makeid, tidyup = TRUE)
}
\arguments{
\item{tidyup}{simplifies json response into a dataframe, Defaults to TRUE}

\item{vin}{makeid in quotes}
}
\description{
vin_dieseltype returns vehicle types associated with a makeid. More information on the NHTSA API is available at https://vpic.nhtsa.dot.gov/api/#
}
\examples{
vin_dieseltype("18697", tidyup = TRUE)
}
\keyword{vin}

