% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieving_functions.R
\name{vin_thief}
\alias{vin_thief}
\title{vin_thief: return a list of Year, Make, Model based on VIN}
\usage{
vin_thief(vin, sec = 5, vector.transform = T, sep = "-")
}
\arguments{
\item{vin}{VIN number in quotes}

\item{sec}{denotes Sys.sleep() between requests if there are multiple, Defaults to 5 seconds}

\item{vector.transform}{Denotes a transformation to a character vector for unique id usage. Defaults to FALSE.}

\item{sep}{symbol denoting separator, can be used to return as a character vector separated by symbols e.g ("model-make-year"). Defaults to "-".}
}
\description{
String returned is useful for feeding into other api's that look up on model and year (FuelEconomy.gov) or rOpenGov/mpg
}
\examples{
vin_thief("1G2HX54K724118697", vector.transform = TRUE, sep = ",")
}
\keyword{vin}

